# Conflicts:
#	app/src/main/java/com/sagrd/travelmanagement/adapters/ClientesAdapter.kt
#	app/src/main/java/com/sagrd/travelmanagement/ui/ClientesFragment.kt
#	app/src/main/res/layout/facturas_pendiente_fragment.xml
#	app/src/main/res/navigation/nav_graph.xml

+++ app/src/main/java/com/sagrd/travelmanagement/model/Cliente.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Clientes(
     @PrimaryKey(autoGenerate = true)
-    val clienteId :Long,
-    var nombres: String,
-    var balance :Double
+    val clienteId : Long,
+    var nombres   : String,
+    var balance   : Float
 )
